Update to rollup 4
