+++ package.json
@@ -49,11 +49,15 @@
   },
   "homepage": "https://github.com/ProsperWorks/copper-sdk#readme",
   "devDependencies": {
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-replace": "^5.0.5",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^11.1.6",
     "@types/chai": "^4.1.7",
     "@types/mocha": "^8.2.2",
-    "@types/sinon": "^7.0.3",
+    "@types/sinon": "^17.0.3",
     "chai": "^4.1.2",
-    "concurrently": "^6.5.1",
+    "concurrently": "^8.2.2",
     "cross-env": "^7.0.3",
     "gitbook-cli": "^2.3.2",
     "gitbook-plugin-anker-enable": "0.0.4",
@@ -71,15 +75,12 @@
     "karma-rollup-preprocessor": "^7.0.8",
     "karma-sinon": "^1.0.5",
     "mocha": "^8.4.0",
-    "rimraf": "^3.0.2",
-    "rollup": "^1.32.1",
-    "rollup-plugin-istanbul": "^3.0.0",
-    "rollup-plugin-json": "^4.0.0",
-    "rollup-plugin-replace": "^2.0.0",
-    "rollup-plugin-typescript2": "^0.30.0",
-    "rollup-plugin-uglify": "^6.0.1",
-    "sinon": "^13.0.2",
-    "tslint": "^5.12.1",
-    "typescript": "^3.2.2"
+    "rimraf": "^5.0.5",
+    "rollup": "^4.17.2",
+    "rollup-plugin-istanbul": "^5.0.0",
+    "sinon": "^17.0.1",
+    "tslib": "^2.6.2",
+    "tslint": "^6.1.3",
+    "typescript": "^5.4.5"
   }
 }
