+++ test/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "../tsconfig",
+  "compilerOptions": {
+    "declaration": false
+  },
   "include": [
     "**/*.spec.ts"
   ]
